pyproject.toml
[tool.poetry]
name = "game-gwent"
version = "0.1.0"
description = ""
authors = ["Asgef <asgefes1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
django-bootstrap5 = "^24.2"
python-dotenv = "^1.0.1"
whitenoise = "^6.6.0"
dj-database-url = "^2.2.0"
gunicorn = "^22.0.0"
rollbar = "^1.0.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
firebase-admin = "^6.5.0"
django-storages = "^1.14.3"
django-extensions = "^3.2.3"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.40"
dadata = "^21.10.1"
requests = "^2.32.3"
yookassa = "^3.3.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
ruff = "^0.4.6"
bpython = "^0.24"
ngrok = "^1.3.0"
pyngrok = "^7.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# Configuring the Ruff tool for source code analysis with options for line length and fixes.
[tool.ruff]
line-length = 80
show-fixes = true
src = ["src"]

select = [
  "E",
  "F",
  "UP",
  "D"
]
